:sparkles: Implement structure to hold cookie data
Instead of writing cookie data straight into the state, create a
structure instead that holds not only said data, but also any unsigning
exception that could've occurred. This is very useful for certain
checks in implementations, as shown in the examples.

BREAKING CHANGE: This change implies rewriting any implementation,
regarding the way they access state data. It's a single line change, so
it's not that bad.